+++ app/controllers/sale_controller.rb
@@ -1,6 +1,11 @@
 class SaleController < ApplicationController
   layout 'sale'
+  layout 'sale_html', only: 'dashboard'
   def index
+    @products = Product.by_store(current_store.id)
+  end
+
+  def dashboard
     @products = Product.by_store(current_store.id)
   end
 end