+++ app/javascript/packs/components.js
@@ -1,0 +1,6 @@
+import Vue from 'vue'
+
+import 'vue-loaders/dist/vue-loaders.css';
+import BaseLoading from './components/BaseLoading.vue'
+
+Vue.component('BaseLoading', BaseLoading)