+++ app/javascript/packs/components/ProductCartList.vue
@@ -1,0 +1,13 @@
+<template>
+    
+</template>
+
+<script>
+    export default {
+        name: "ProductCartList"
+    }
+</script>
+
+<style scoped>
+
+</style>