+++ app/javascript/packs/components/SaleHeader.vue
@@ -1,0 +1,90 @@
+<template>
+    <nav class="navbar navbar-expand-lg navbar-dark bg-park">
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+
+        <div class="collapse navbar-collapse" id="navbarSupportedContent">
+            <form class="form col-md-5">
+                <div class="input-group">
+                    <div class="input-group-prepend">
+                      <span class="input-group-text bg-white border-right-0">
+                        <i class="fa fa-search"></i>
+                      </span>
+                    </div>
+                    <input type="text" class="form-control form-control-sm border-left-0" placeholder="Nhập mã hoặc tên mặt hàng (F3)" aria-label="Tìm mặt hàng" id="search-product-input" autofocus>
+                </div>
+            </form>
+
+            <ul class="navbar-nav mx-auto">
+                <div class="d-flex justify-content-center help-text-login">
+                    <a class="help-text-login" @click="openReport">Báo cáo</a>
+                    <span class="help-text-login">|</span>
+                    <a class="help-text-login" @click="openManage">Quản lý</a>
+                    <span class="help-text-login">|</span>
+                    <a href="#" class="help-text-login">Nhập trả</a>
+                </div>
+            </ul>
+
+        </div>
+
+        <div>
+            <button type="button" class="btn btn-link text-light" @click="toggleFullScreen"><i class="fa fa-arrows-alt" id="btn-fullscreen-sale"></i></button>
+            <button type="button" class="btn btn-link text-light"><i class="fa fa-question"></i></button>
+        </div>
+
+        <div class="navbar-select-branch">
+
+        </div>
+
+        <ul class="navbar-nav">
+            <li class="nav-item dropdown">
+                <a class="nav-link dropdown-toggle navbar-user" href="#" id="navbarDropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false" v-html="welcome_user"
+                   @click="toggleUserMenu" @blur="closeUserMenu">
+                    <br>
+                </a>
+                <div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarDropdownMenuLink" :class="{'show' : showUserMenu}">
+                    <a href="#" class="dropdown-item">Tài khoản</a>
+                    <div class="dropdown-divider"></div>
+                    <a href="#" class="dropdown-item">Đăng xuất</a>
+                </div>
+            </li>
+        </ul>
+
+    </nav>
+</template>
+
+<script>
+    export default {
+        data: () => ({
+            showUserMenu: false
+        }),
+        props: ['user'],
+        computed: {
+            welcome_user () {
+                return "Xin chào<br>" + this.user.fullname
+            }
+        },
+        methods: {
+            openReport () {
+                window.open('/manage/reports', '_blank')
+            },
+            openManage () {
+                window.open('/manage', '_blank')
+            },
+            toggleFullScreen () {
+                // Should write by state of VueJS
+            },
+            toggleUserMenu () {
+                this.showUserMenu = !this.showUserMenu
+            },
+            closeUserMenu () {
+                this.showUserMenu = false
+            }
+        }
+    }
+</script>
+
+<style scoped>
+
+</style>