+++ app/javascript/packs/components/SaleUser.vue
@@ -1,0 +1,20 @@
+<template>
+    <div class="sale-user d-flex mt-2">
+        <div class="mr-auto">
+
+        </div>
+        <div>
+
+        </div>
+    </div>
+</template>
+
+<script>
+    export default {
+        name: "SaleUser"
+    }
+</script>
+
+<style scoped>
+
+</style>