+++ app/javascript/packs/state.js
@@ -1,0 +1,4 @@
+export default {
+    user: null,
+    products: null
+}