+++ config/initializers/content_security_policy.rb
@@ -15,7 +15,7 @@
 #   # Specify URI for violation reports
 #   # policy.report_uri "/csp-violation-report-endpoint"
    if Rails.env.development?
-     policy.script_src :self, :https, :unsafe_eval
+     policy.script_src :self, :https, :unsafe_eval, :unsafe_inline
    else
      policy.script_src :self, :https
    end